+++ R/macro-bookdown.R
@@ -20,12 +20,7 @@ NULL
 #' @inheritParams step_do_push_deploy
 #' @inheritParams step_install_pkg
 #' @param ... Passed on to [step_build_bookdown()]
-#' @param travis_private_key_name `string`\cr
-#'   Only needed when deploying from builds on Travis CI.
-#'   If you have set a custom name for the private key during creation of the
-#'   SSH key pair in [travis::use_travis_deploy()] or via [use_tic], you need
-#'   to pass this name here. If not set, `"TRAVIS_DEPLOY_KEY"` will be used
-#'   by default.
+#' @template private_key_name
 #' @family macros
 #' @export
 #' @examples
@@ -46,9 +41,7 @@ do_bookdown <- function(...,
                         remote_url = NULL,
                         commit_message = NULL,
                         commit_paths = ".",
-                        travis_private_key_name = "TRAVIS_DEPLOY_KEY") {
-
-  name <- travis_private_key_name
+                        private_key_name = "TIC_DEPLOY_KEY") {
 
   #' @param deploy `[flag]`\cr
   #'   If `TRUE`, deployment setup is performed
@@ -61,8 +54,11 @@ do_bookdown <- function(...,
     #'
     #'   1. The repo can be pushed to (see [ci_can_push()]).
     # account for old default "id_rsa"
-    cli_text("Using {name} env var as the private key name for SSH deployment.")
-    deploy <- ci_can_push(name = name)
+    private_key_name <- private_key_name
+
+    cli_text("Using {private_key_name} env var as the private key name for
+             SSH deployment.", wrap = TRUE)
+    deploy <- ci_can_push(private_key_name = private_key_name)
 
     #'   2. The `branch` argument is `NULL`
     #'   (i.e., if the deployment happens to the active branch),
@@ -84,9 +80,9 @@ do_bookdown <- function(...,
     #' 1. [step_setup_push_deploy()] in the `"before_deploy"` stage
     #'    (if `deploy` is set),
     #'
-    name <- travis_private_key_name
+    private_key_name <- private_key_name
     get_stage("before_deploy") %>%
-      add_step(step_setup_ssh(name = name)) %>%
+      add_step(step_setup_ssh(private_key_name = !!private_key_name)) %>%
       add_step(step_setup_push_deploy(
         path = !!enquo(path),
         branch = !!enquo(branch),