+++ R/macro-package-checks.R
@@ -35,8 +35,10 @@ do_package_checks <- function(...,
                               args = NULL,
                               build_args = NULL,
                               error_on = "warning",
-                              repos = repo_default(), timeout = Inf,
-                              type = NULL, check_dir = NULL) {
+                              repos = repo_default(),
+                              timeout = Inf,
+                              type = getOption("pkgType"),
+                              check_dir = NULL) {
   #' @description
   #' 1. [step_install_deps()] in the `"install"` stage, using the
   #'    `repos` argument.