+++ R/steps-ssh.R
@@ -1,3 +1,5 @@
+# AddToKnownHosts --------------------------------------------------------------
+
 # This code can only run as part of a CI run
 # nocov start
 
@@ -11,7 +13,8 @@ AddToKnownHosts <- R6Class(
     },
 
     run = function() {
-      message("Running ssh-keyscan for ", private$host)
+      cli_text("{.fun step_add_to_known_hosts}: Running ssh-keyscan for
+               {private$host}.")
       keyscan_result <- system2(
         "ssh-keyscan",
         c("-H", shQuote(private$host)),
@@ -24,7 +27,7 @@ AddToKnownHosts <- R6Class(
         dirname(known_hosts_path),
         showWarnings = FALSE, recursive = TRUE
       )
-      message("Adding to ", known_hosts_path)
+      cli_text("Adding to {known_hosts_path}.")
       write(keyscan_result, known_hosts_path, append = TRUE)
     },
 
@@ -61,41 +64,50 @@ step_add_to_known_hosts <- function(host = "github.com") {
   AddToKnownHosts$new(host = host)
 }
 
+# InstallSSHKeys ---------------------------------------------------------------
+
 InstallSSHKeys <- R6Class(
   "InstallSSHKeys",
   inherit = TicStep,
 
   public = list(
-    initialize = function(name = "TRAVIS_DEPLOY_KEY") {
+    initialize = function(private_key_name = "TIC_DEPLOY_KEY") {
       # for backward comp, if "id_rsa" exists we take this key
-      private$name <- compat_ssh_key(name = name)
+      private$private_key_name <- compat_ssh_key(private_key_name = private_key_name)
     },
 
     run = function() {
-      name <- private$name
+      private_key_name <- private$private_key_name
 
-      deploy_key_path <- file.path("~", ".ssh", name)
+      deploy_key_path <- file.path("~", ".ssh", private_key_name)
       dir.create(
         dirname(deploy_key_path),
         recursive = TRUE, showWarnings = FALSE
       )
-      message("Writing deploy key to ", deploy_key_path)
+      cli_text("{.fun step_install_ssh_keys}:
+               Writing deploy key to {.file {deploy_key_path}}.",
+        wrap = TRUE
+      )
       if (file.exists(deploy_key_path)) {
-        message("Not overwriting existing SSH key.")
+        cli_text("Not overwriting existing SSH key.")
         return()
       }
       writeLines(
-        rawToChar(openssl::base64_decode(Sys.getenv(name))),
+        rawToChar(openssl::base64_decode(Sys.getenv(private_key_name))),
         deploy_key_path
       )
 
-      Sys.chmod(file.path("~", ".ssh", name), "600")
+      Sys.chmod(file.path("~", ".ssh", private_key_name), "600")
 
       # set the ssh command which which git should use including the key name
       git2r::config(
         core.sshCommand = sprintf(
-          "ssh -i ~/.ssh/%s -F /dev/null",
-          name
+          paste0(
+            "ssh ",
+            "-i ~/.ssh/%s -F /dev/null ",
+            "-o LogLevel=error"
+          ),
+          private_key_name
         ),
         global = TRUE
       )
@@ -107,16 +119,19 @@ InstallSSHKeys <- R6Class(
 
     check = function() {
 
-      # only if non-interactive and TRAVIS_DEPLOY_KEY env var is available
+      # only if non-interactive and TIC_DEPLOY_KEY env var is available
       if (!ci_is_interactive()) {
-        if (!ci_can_push(private$name)) {
-          cli_alert_danger("Deployment was requested but the build is not able to
-                       deploy. We checked for env var {.var {private$name}} but could
-                       not find it as an env var in the current build.
-                       Double-check if it exists. Calling
-                       {.fun travis::use_travis_deploy} may help resolving
-                       issues.", wrap = TRUE)
-          stopc("This build cannot deploy to Github.")
+        if (!ci_can_push(private$private_key_name)) {
+          cli_alert_danger("{.fun step_install_ssh_keys}: Deployment was
+          requested but the build is not able to deploy.
+          We checked for env var {.var {private$private_key_name}}
+          but could not find it as an env var in the current build.
+          Double-check if it exists.
+          Calling {.fun travis::use_travis_deploy} or
+          {.fun tic::use_ghactions_deploy} may help resolving issues.",
+            wrap = TRUE
+          )
+          stopc("This build cannot deploy to GitHub.")
         }
         TRUE
       } else {
@@ -126,7 +141,7 @@ InstallSSHKeys <- R6Class(
   ),
 
   private = list(
-    name = NULL
+    private_key_name = NULL
   )
 )
 
@@ -136,15 +151,14 @@ InstallSSHKeys <- R6Class(
 #' to a file in `~/.ssh`.
 #' Only run in non-interactive settings and if the environment variable
 #' exists and is non-empty.
-#' The [travis::use_travis_deploy()] and [use_tic()] functions encode a private
-#' key as an environment variable for use with this function.
+#' [travis::use_travis_deploy()] , [use_ghactions_deploy()] and [use_tic()]
+#' functions encode a private key as an environment variable for use with this
+#' function.
 #'
-#' @param name `[string]`\cr
-#'   Name of the environment variable and the target file, default:
-#'   `"TRAVIS_DEPLOY_KEY"`.
+#' @template private_key_name
 #'
 #' @family steps
-#' @seealso [travis::use_travis_deploy()], [use_tic()]
+#' @seealso [travis::use_travis_deploy()], [use_tic()], [use_ghactions_deploy()]
 #' @export
 #' @examples
 #' dsl_init()
@@ -153,10 +167,12 @@ InstallSSHKeys <- R6Class(
 #'   add_step(step_install_ssh_keys())
 #'
 #' dsl_get()
-step_install_ssh_keys <- function(name = "TRAVIS_DEPLOY_KEY") {
-  name <- compat_ssh_key(name = name)
-  InstallSSHKeys$new(name = name)
+step_install_ssh_keys <- function(private_key_name = "TIC_DEPLOY_KEY") {
+  private_key_name <- compat_ssh_key(private_key_name = private_key_name)
+  InstallSSHKeys$new(private_key_name = private_key_name)
 }
+
+# TestSSH ----------------------------------------------------------------------
 
 TestSSH <- R6Class(
   "TestSSH",
@@ -164,23 +180,24 @@ TestSSH <- R6Class(
 
   public = list(
     initialize = function(url = "git@github.com",
-                          verbose = "-v",
-                          name = "TRAVIS_DEPLOY_KEY") {
+                          verbose = "",
+                          private_key_name = "TIC_DEPLOY_KEY") {
       private$url <- url
       private$verbose <- verbose
-      private$name <- name
+      private$private_key_name <- private_key_name
     },
 
     run = function() {
 
-      message("Trying to ssh into ", private$url)
-      message("Using command: '", sprintf(
-        "ssh -i %s %s %s'",
-        file.path("~", ".ssh", private$name),
-        private$url, private$verbose
-      ))
+      cli_text("{.fun step_test_ssh}: Trying to ssh into {private$url}")
+      cli_text("{.fun step_test_ssh}: Using command:
+               {.code ssh -i ~/.ssh/{private$private_key_name}
+               -o LogLevel=error
+               {private$url} {private$verbose}}", wrap = TRUE)
+      # suppress the warning about adding the IP to .ssh/known_hosts
       system2("ssh", c(
-        "-i", file.path("~", ".ssh", private$name),
+        "-o", "LogLevel=error",
+        "-i", file.path("~", ".ssh", private$private_key_name),
         private$url, private$verbose
       ))
     }
@@ -189,7 +206,7 @@ TestSSH <- R6Class(
   private = list(
     url = NULL,
     verbose = NULL,
-    name = NULL
+    private_key_name = NULL
   )
 )
 
@@ -203,7 +220,7 @@ TestSSH <- R6Class(
 #' @param url `[string]`\cr
 #'   URL to establish SSH connection with, by default `git@github.com`
 #' @param verbose `[string]`\cr
-#'   Verbosity, by default `"-v"`. Use `"-vvv"` for more verbosity.
+#'   Verbosity, by default `""`. Use `-v` or `"-vvv"` for more verbosity.
 #' @inheritParams step_install_ssh_keys
 #' @family steps
 #' @export
@@ -215,24 +232,28 @@ TestSSH <- R6Class(
 #'
 #' dsl_get()
 step_test_ssh <- function(url = "git@github.com",
-                          verbose = "-v",
-                          name = "TRAVIS_DEPLOY_KEY") {
-  TestSSH$new(url = url, verbose = verbose, name = name)
+                          verbose = "",
+                          private_key_name = "TIC_DEPLOY_KEY") {
+  TestSSH$new(url = url, verbose = verbose, private_key_name = private_key_name)
 }
+
+# SetupSSH ---------------------------------------------------------------------
 
 SetupSSH <- R6Class(
   "SetupSSH",
   inherit = TicStep,
 
   public = list(
-    initialize = function(name = "TRAVIS_DEPLOY_KEY", host = "github.com",
-                          url = paste0("git@", host), verbose = "-v") {
+    initialize = function(private_key_name = "TIC_DEPLOY_KEY",
+                          host = "github.com",
+                          url = paste0("git@", host),
+                          verbose = "") {
 
-      private$install_ssh_keys <- step_install_ssh_keys(name = name)
+      private$install_ssh_keys <- step_install_ssh_keys(private_key_name = private_key_name)
       private$add_to_known_hosts <- step_add_to_known_hosts(host = host)
       private$test_ssh <- step_test_ssh(
         url = url, verbose = verbose,
-        name = name
+        private_key_name = private_key_name
       )
     },
 
@@ -261,7 +282,8 @@ SetupSSH <- R6Class(
         cli_alert_info("{.fun SetupSSH$check}: {.fun test_ssh} failed.")
         return(FALSE)
       }
-      TRUE
+      cli_alert_success("{.fun step_setup_ssh} Everything ok.")
+      return(TRUE)
     }
   ),
 
@@ -294,17 +316,25 @@ SetupSSH <- R6Class(
 #'   add_step(step_setup_ssh(host = "gitlab.com"))
 #'
 #' dsl_get()
-step_setup_ssh <- function(name = "TRAVIS_DEPLOY_KEY", host = "github.com",
-                           url = paste0("git@", host), verbose = "-v") {
-  SetupSSH$new(name = name, host = host, url = url, verbose = verbose)
+step_setup_ssh <- function(private_key_name = "TIC_DEPLOY_KEY",
+                           host = "github.com",
+                           url = paste0("git@", host),
+                           verbose = "") {
+  SetupSSH$new(
+    private_key_name = private_key_name, host = host,
+    url = url, verbose = verbose
+  )
 }
 
-compat_ssh_key <- function(name) {
-  # for backward comp, if "id_rsa" exists we take this key
-  if (ci_has_env("id_rsa") && !ci_has_env(name)) {
-    name <- "id_rsa"
+compat_ssh_key <- function(private_key_name) {
+  # for backward comp, if "id_rsa" or "TRAVIS_DEPLOY_KEY" exists we take this
+  # key
+  if (ci_has_env("id_rsa") && !ci_has_env(private_key_name)) {
+    private_key_name <- "id_rsa"
+  } else if (ci_has_env("TRAVIS_DEPLOY_KEY") && !ci_has_env(private_key_name)) {
+    private_key_name <- "TRAVIS_DEPLOY_KEY"
   }
-  name
+  private_key_name
 }
 
 # This code can only run as part of a CI run